Add default value to faulty message id
The message id "Successfully delete items" is erroneous; it should read
"Successfully deleted items".

Ideally the message id would be changed.  While this is possible and
probably should be done in the future, this imposes problems, because
the source of that id is another package; we would impose a new version
dependency.

Thus, what we can do immediately, is to apply a default value which will
be used for all languages which lack an own translation, and act as a
hint to translators.