Updated Language-Codes in po file headers.
These headers are not used in Plone to determine the language: that is
done by inspecting the directory name.  But the i18ndude script uses
the Language-Code header when printing statistics.  Several were set
to en or to for example zh_CN (as the directory name should be)
instead of zh-cn (as the language code should be).